fil-devops-flask/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
snowflake-connector-python
snowflake-sqlalchemy
pytest
pytest-flask
gunicorn